requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.0.6
django-celery-beat==1.6.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-timezone-field==3.1
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.7
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
zipp==3.0.0
